pyproject.toml
[build-system]
requires = ["setuptools>=70.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-api"
version = "2.0.0"
description = "Retrieval-Augmented Generation API with Elasticsearch and multiple LLM backends."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "httpx>=0.27.0",
    "openai>=1.35.0",
    "elasticsearch>=8.14.0",
    "sentence-transformers>=3.0.0",
    "loguru>=0.7.0",
    "pypdf>=4.0.0",
    "pandas>=2.0.0",
    "python-docx>=1.1.0",
    "llama-index-core>=0.11.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
